# [SPARK-46228][SQL] Insert window group limit node for limit outside of window