--- a/sql/core/src/test/scala/org/apache/spark/sql/DataFrameWindowFunctionsSuite.scala
+++ b/sql/core/src/test/scala/org/apache/spark/sql/DataFrameWindowFunctionsSuite.scala
@@ -1276,6 +1276,141 @@ class DataFrameWindowFunctionsSuite extends QueryTest
     )
   }
 
+  test("SPARK-46228: Insert window group limit node for limit outside of window") {
+
+    val nullStr: String = null
+    val df = Seq(
+      ("a", 0, "c"),
+      ("a", 1, "x"),
+      ("a", 2, "y"),
+      ("a", 3, "z"),
+      ("a", 4, ""),
+      ("a", 4, ""),
+      ("b", 1, "h"),
+      ("b", 1, "n"),
+      ("c", 1, "z"),
+      ("c", 1, "a"),
+      ("c", 2, nullStr)).toDF("key", "value", "order")
+    val window = Window.partitionBy($"key").orderBy($"order".asc_nulls_first)
+    val window2 = Window.partitionBy($"key").orderBy($"order".desc_nulls_first)
+    val window3 = Window.orderBy($"order".asc_nulls_first)
+    val rowsWindow = window.rowsBetween(Window.unboundedPreceding, Window.currentRow)
+    val rangeWindow = window.rangeBetween(Window.unboundedPreceding, Window.currentRow)
+    val rowsWindow2 = window2.rowsBetween(Window.unboundedPreceding, Window.currentRow)
+    val rangeWindow2 = window2.rangeBetween(Window.unboundedPreceding, Window.currentRow)
+    val rowsWindow3 = window3.rowsBetween(Window.unboundedPreceding, Window.currentRow)
+    val rangeWindow3 = window3.rangeBetween(Window.unboundedPreceding, Window.currentRow)
+
+    Seq(-1, 100).foreach { threshold =>
+      withSQLConf(SQLConf.WINDOW_GROUP_LIMIT_THRESHOLD.key -> threshold.toString) {
+        // RowFrame
+        checkAnswer(df.withColumn("rn", row_number().over(window)).limit(1),
+          Seq(
+            Row("a", 4, "", 1)
+          )
+        )
+        checkAnswer(df.withColumn("rn", rank().over(window2)).limit(7),
+          Seq(
+            Row("a", 0, "c", 4),
+            Row("a", 1, "x", 3),
+            Row("a", 2, "y", 2),
+            Row("a", 3, "z", 1),
+            Row("a", 4, "", 5),
+            Row("a", 4, "", 5),
+            Row("b", 1, "n", 1)
+          )
+        )
+        checkAnswer(df.withColumn("rn", dense_rank().over(window3)).limit(11),
+          Seq(
+            Row("a", 0, "c", 4),
+            Row("a", 1, "x", 7),
+            Row("a", 2, "y", 8),
+            Row("a", 3, "z", 9),
+            Row("a", 4, "", 2),
+            Row("a", 4, "", 2),
+            Row("b", 1, "h", 5),
+            Row("b", 1, "n", 6),
+            Row("c", 1, "a", 3),
+            Row("c", 1, "z", 9),
+            Row("c", 2, nullStr, 1)
+          )
+        )
+
+        // RangeFrame
+        checkAnswer(df.withColumn("sum_value", sum("value").over(window)).limit(1),
+          Seq(
+            Row("a", 4, "", 8)
+          )
+        )
+        checkAnswer(df.withColumn("sum_value", sum("value").over(window2)).limit(7),
+          Seq(
+            Row("a", 0, "c", 6),
+            Row("a", 1, "x", 6),
+            Row("a", 2, "y", 5),
+            Row("a", 3, "z", 3),
+            Row("a", 4, "", 14),
+            Row("a", 4, "", 14),
+            Row("b", 1, "n", 1)
+          )
+        )
+        checkAnswer(df.withColumn("sum_value", sum("value").over(window3)).limit(11),
+          Seq(
+            Row("a", 0, "c", 11),
+            Row("a", 1, "x", 14),
+            Row("a", 2, "y", 16),
+            Row("a", 3, "z", 20),
+            Row("a", 4, "", 10),
+            Row("a", 4, "", 10),
+            Row("b", 1, "h", 12),
+            Row("b", 1, "n", 13),
+            Row("c", 1, "a", 11),
+            Row("c", 1, "z", 20),
+            Row("c", 2, nullStr, 2)
+          )
+        )
+
+        // Both RowFrame and RangeFrame exist
+        checkAnswer(
+          df.withColumn("sum_value1", sum("value").over(rowsWindow))
+            .withColumn("sum_value2", sum("value").over(rangeWindow)).limit(1),
+          Seq(
+            Row("a", 4, "", 4, 8)
+          )
+        )
+        checkAnswer(
+          df.withColumn("sum_value1", sum("value").over(rowsWindow2))
+            .withColumn("sum_value2", sum("value").over(rangeWindow2)).limit(7),
+          Seq(
+            Row("a", 0, "c", 6, 6),
+            Row("a", 1, "x", 6, 6),
+            Row("a", 2, "y", 5, 5),
+            Row("a", 3, "z", 3, 3),
+            Row("a", 4, "", 10, 14),
+            Row("a", 4, "", 14, 14),
+            Row("b", 1, "n", 1, 1)
+          )
+        )
+        checkAnswer(
+          df.withColumn("sum_value1", sum("value").over(rowsWindow3))
+            .withColumn("sum_value2", sum("value").over(rangeWindow3)).limit(11),
+          Seq(
+            Row("a", 0, "c", 11, 11),
+            Row("a", 1, "x", 14, 14),
+            Row("a", 2, "y", 16, 16),
+            Row("a", 3, "z", 19, 20),
+            Row("a", 4, "", 10, 10),
+            Row("a", 4, "", 6, 10),
+            Row("b", 1, "h", 12, 12),
+            Row("b", 1, "n", 13, 13),
+            Row("c", 1, "a", 11, 11),
+            Row("c", 1, "z", 20, 20),
+            Row("c", 2, nullStr, 2, 2)
+          )
+        )
+      }
+    }
+  }
+
   test("SPARK-37099: Insert window group limit node for top-k computation") {
 
     val nullStr: String = null