@zml1206 2023-12-06
> Although I understood the mechanism, please add more detail in the PR description. So as other contributors could understand.

Done