@beliefer 2023-12-09
@zml1206 The PR description is not clear enough. Please modify
`We can extract the limit value 2 (("a", 4, ""),("a", 4, "")) on window group a because window expression is RangeFrame and after sort, the sort values ​​of the first row and the second row are the same, limit value 1 (("b", 1, "h")) on window group b in WindowGroupLimitExec.`
You can change to `We infer the rank() on window group a because window expression is RangeFrame and after sort. WindowGroupLimitExec will select `RankLimitIterator` to select the range values. We infer the rownumber() on window group b. `WindowGroupLimitExec` will select `SimpleLimitIterator` to select the row values.`